fix: remove issues with name = null

+++ src/main/java/tool/bico/writer/CommitWriter.java
@@ -7,33 +7,31 @@ import org.springframework.batch.item.ItemWriter;
 import tool.bico.model.Commit;
 import tool.bico.model.CommitIssue;
 import tool.bico.model.service.CommitIssueService;
+import tool.bico.model.service.CommitService;
 
 public class CommitWriter implements ItemWriter<CommitIssue> {
 
+	private CommitService commitService;
 	private CommitIssueService commitIssueService;
 	
-	public CommitWriter(CommitIssueService commitIssueService) {
+	public CommitWriter(CommitService commitService, CommitIssueService commitIssueService) {
+		this.commitService = commitService;
 		this.commitIssueService = commitIssueService;
 	}
 	
 	@Override
 	public void write(List<? extends CommitIssue> items) throws Exception {
-		for(CommitIssue commitIssue : items) {
-			/*if(commitIssue.getName() == null) {
-				//commitIssue.getCommits().clear();
-				System.err.println("trying to delete); remove first issue from commits");
+		for(CommitIssue commitIssue : items) {	
+			if(commitIssue.getName() == null) {
 				for(Commit c : commitIssue.getCommits()) {
 					c.removeCommitIssue(commitIssue);
+					commitService.update(c);
 				}
-				System.err.println("trying to clear commits list in issue");
-				commitIssue.getCommits().clear();
-				System.err.println("start to delete");
+				
 				commitIssueService.delete(commitIssue);
-				System.err.println("DELETE SUCCESSFULL");
 			} else {
 				commitIssueService.update(commitIssue);
-			}*/
-			commitIssueService.update(commitIssue);
+			}
 		}
 	}
 }
